+++ internal/k8s/crdcontroller.go
@@ -118,7 +118,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				}
 				ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
 				if ok && resVer.(string) == hostrule.ResourceVersion {
-					utils.AviLog.Debugf("Same resource version returning")
+					utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
 					return
 				}
 				utils.AviLog.Debugf("key: %s, msg: ADD", key)
@@ -162,6 +162,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				namespace, _, _ := cache.SplitMetaNamespaceKey(utils.ObjKey(hostrule))
 				key := lib.HostRule + "/" + utils.ObjKey(hostrule)
 				utils.AviLog.Debugf("key: %s, msg: DELETE", key)
+				objects.SharedResourceVerInstanceLister().Delete(key)
 				bkt := utils.Bkt(namespace, numWorkers)
 				c.workqueue[bkt].AddRateLimited(key)
 			},
@@ -184,7 +185,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				}
 				ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
 				if ok && resVer.(string) == httprule.ResourceVersion {
-					utils.AviLog.Debugf("Same resource version returning")
+					utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
 					return
 				}
 				utils.AviLog.Debugf("key: %s, msg: ADD", key)
@@ -232,6 +233,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				utils.AviLog.Debugf("key: %s, msg: DELETE", key)
 				// no need to validate for delete handler
 				bkt := utils.Bkt(namespace, numWorkers)
+				objects.SharedResourceVerInstanceLister().Delete(key)
 				c.workqueue[bkt].AddRateLimited(key)
 			},
 		}
@@ -253,7 +255,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				}
 				ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
 				if ok && resVer.(string) == aviinfra.ResourceVersion {
-					utils.AviLog.Debugf("Same resource version returning")
+					utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
 					return
 				}
 				utils.AviLog.Debugf("key: %s, msg: ADD", key)
@@ -297,6 +299,7 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 				key := lib.AviInfraSetting + "/" + utils.ObjKey(aviinfra)
 				namespace, _, _ := cache.SplitMetaNamespaceKey(utils.ObjKey(aviinfra))
 				utils.AviLog.Debugf("key: %s, msg: DELETE", key)
+				objects.SharedResourceVerInstanceLister().Delete(key)
 				// no need to validate for delete handler
 				bkt := utils.Bkt(namespace, numWorkers)
 				c.workqueue[bkt].AddRateLimited(key)
@@ -304,6 +307,13 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 		}
 
 		informer.AviInfraSettingInformer.Informer().AddEventHandler(aviInfraEventHandler)
+	}
+	return
+}
+
+func (c *AviController) AddCrdIndexer() {
+	informer := lib.GetCRDInformers()
+	if lib.GetAviInfraSettingEnabled() {
 		informer.AviInfraSettingInformer.Informer().AddIndexers(
 			cache.Indexers{
 				lib.SeGroupAviSettingIndex: func(obj interface{}) ([]string, error) {
@@ -316,8 +326,6 @@ func (c *AviController) SetupAKOCRDEventHandlers(numWorkers uint32) {
 			},
 		)
 	}
-
-	return
 }
 
 // SetupIstioCRDEventHandlers handles setting up of Istio CRD event handlers
@@ -334,6 +342,11 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			namespace, _, _ := cache.SplitMetaNamespaceKey(utils.ObjKey(vs))
 			key := lib.IstioVirtualService + "/" + utils.ObjKey(vs)
 			utils.AviLog.Debugf("key: %s, msg: ADD", key)
+			ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
+			if ok && resVer.(string) == vs.ResourceVersion {
+				utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
+				return
+			}
 			bkt := utils.Bkt(namespace, numWorkers)
 			c.workqueue[bkt].AddRateLimited(key)
 		},
@@ -372,6 +385,7 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			key := lib.IstioVirtualService + "/" + utils.ObjKey(vs)
 			utils.AviLog.Debugf("key: %s, msg: DELETE", key)
 			bkt := utils.Bkt(namespace, numWorkers)
+			objects.SharedResourceVerInstanceLister().Delete(key)
 			c.workqueue[bkt].AddRateLimited(key)
 		},
 	}
@@ -388,6 +402,11 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			key := lib.IstioDestinationRule + "/" + utils.ObjKey(dr)
 			utils.AviLog.Debugf("key: %s, msg: ADD", key)
 			bkt := utils.Bkt(namespace, numWorkers)
+			ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
+			if ok && resVer.(string) == dr.ResourceVersion {
+				utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
+				return
+			}
 			c.workqueue[bkt].AddRateLimited(key)
 		},
 		UpdateFunc: func(old, new interface{}) {
@@ -425,6 +444,7 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			key := lib.IstioDestinationRule + "/" + utils.ObjKey(dr)
 			utils.AviLog.Debugf("key: %s, msg: DELETE", key)
 			bkt := utils.Bkt(namespace, numWorkers)
+			objects.SharedResourceVerInstanceLister().Delete(key)
 			c.workqueue[bkt].AddRateLimited(key)
 		},
 	}
@@ -440,6 +460,11 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			namespace, _, _ := cache.SplitMetaNamespaceKey(utils.ObjKey(vs))
 			key := lib.IstioGateway + "/" + utils.ObjKey(vs)
 			utils.AviLog.Debugf("key: %s, msg: ADD", key)
+			ok, resVer := objects.SharedResourceVerInstanceLister().Get(key)
+			if ok && resVer.(string) == vs.ResourceVersion {
+				utils.AviLog.Debugf("key: %s, msg: Same resource version returning", key)
+				return
+			}
 			bkt := utils.Bkt(namespace, numWorkers)
 			c.workqueue[bkt].AddRateLimited(key)
 		},
@@ -478,6 +503,7 @@ func (c *AviController) SetupIstioCRDEventHandlers(numWorkers uint32) {
 			key := lib.IstioGateway + "/" + utils.ObjKey(vs)
 			utils.AviLog.Debugf("key: %s, msg: DELETE", key)
 			bkt := utils.Bkt(namespace, numWorkers)
+			objects.SharedResourceVerInstanceLister().Delete(key)
 			c.workqueue[bkt].AddRateLimited(key)
 		},
 	}
